functions/next-song-updater/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-cloud-pubsub==2.*
redis==5.0.8
